refactor(HttpRequestHandler): reuse single stateless HttpRequestHandler instance
